Moved run_with_coverage into pytest-shutil.run
